+++ src/main/java/dev/sidgames/wmo/WayMoreOres.java
@@ -1,7 +1,7 @@
 package dev.sidgames.wmo;
 
-import dev.sidgames.wmo.block.Blocks;
-import dev.sidgames.wmo.item.Items;
+import dev.sidgames.wmo.block.ModBlocks;
+import dev.sidgames.wmo.item.ModItems;
 import dev.sidgames.wmo.util.WMORegistries;
 import dev.sidgames.wmo.world.feature.WMOConfiguredFeaturesHelper;
 import dev.sidgames.wmo.world.gen.WMOWorldGenHelper;
@@ -10,8 +10,8 @@ import net.fabricmc.api.ModInitializer;
 public class WayMoreOres implements ModInitializer {
 	@Override
 	public void onInitialize() {
-		Items.registerItems();
-		Blocks.registerBlocks();
+		ModItems.registerItems();
+		ModBlocks.registerBlocks();
 		WMORegistries.registerModStuffs();
 		WMOConfiguredFeaturesHelper.registerConfiguredFeatures();
 		WMOWorldGenHelper.generateModWorldGen();